+++ VerticalTimeFading/transact_set.hpp
@@ -19,7 +19,7 @@ public:
 	double sum( const vert::fade_set &fade ) const;
 	double sum() const;
 	std::size_t size() const;
-	bool find( int32_t value ) const;
+	bool find( size_t value ) const;
 	int operator[]( std::size_t index ) const;
 	std::string pretty() const;
